WIP Upgrade ic-pos
