+++ motoko/ic-pos/.gitignore
@@ -32,3 +32,5 @@ yarn-error.log*
 
 # misc
 scratch/
+canister_ids.json
+
