+++ motoko/ic-pos/README.md
@@ -35,12 +35,12 @@ https://github.com/kristoferlund/ic-pos/assets/9698363/f67d9952-3ee1-4876-a5e5-6
 
 ### Backend
 
-The backend is written in Motoko and consists of one canister, `icpos`. It exposes four public methods:
+The backend is written in Motoko and consists of one canister, `icpos`. It exposes five public methods:
 
 - `getMerchant` - returns the store configuration for a given principal.
 - `updateMerchant` - updates the store configuration for a given principal.
 - `setCourierApiKey` - sets the Courier API key used to send email and SMS notifications. Only the canister controller can call this method.
-- `setLedgerId` - sets the ledger ID to monitor for transactions. Only the canister controller can call this method.
+- `setLedgerId` - xsets the ledger ID to monitor for transactions. Only the canister controller can call this method.
 - `getLogs` - The canister maintains a debug log that can be fetched using this method.
 
 In addition to the public methods, the canister uses a [timer](https://internetcomputer.org/docs/current/motoko/main/timers/) to monitor ledger transactions. When a new transaction is found that matches a configured store – depending on the store settings – the canister will send a notification either by email or SMS.
@@ -66,6 +66,8 @@ The frontend interacts with the following IC canisters:
 
 ### Step 1: Start a local instance of the replica:
 
+(To complete steps 1-7 in one go, you can run the `./scripts/deploy.sh` script.)
+
 ```bash
 dfx start --clean --background
 ```
@@ -75,7 +77,7 @@ dfx start --clean --background
 Integration with the [Internet Identity](https://internetcomputer.org/internet-identity/) allows store owners to securely setup and manage their store. The Internet Identity canister is already deployed on the IC mainnet. For local development, you need to deploy it to your local instance of the IC.
 
 ```bash
-dfx deploy --network local internet_identity
+dfx deploy internet_identity
 ```
 
 ### Step 3: Save the current principal as a variable:
@@ -101,7 +103,7 @@ Take a moment to read the details of the call we are making below. Not only are 
 - Setting the transfer fee to 10 LCKBTC.
 
 ```bash
-dfx deploy --network local --specified-id mxzaz-hqaaa-aaaar-qaada-cai icrc1_ledger --argument '
+dfx deploy icrc1_ledger --argument '
   (variant {
     Init = record {
       token_name = "Local ckBTC";
@@ -134,7 +136,7 @@ dfx deploy --network local --specified-id mxzaz-hqaaa-aaaar-qaada-cai icrc1_ledg
 The index canister syncs the ledger transactions and indexes them by account.
 
 ```bash
-dfx deploy --network local icrc1_index --argument '
+dfx deploy icrc1_index --argument '
   record {
    ledger_id = (principal "mxzaz-hqaaa-aaaar-qaada-cai");
   }
@@ -148,7 +150,7 @@ The icpos canister manages the store configuration and sends notifications when 
 The `--argument '(0)'` argument is used to initialize the canister with `startBlock` set to 0. This is used to tell the canister to start monitoring the ledger from block 0. When deploying to the IC mainnet, this should be set to the current block height to prevent the canister from processing old transactions.
 
 ```bash
-dfx deploy --network local icpos --argument '(0)'
+dfx deploy icpos --argument '(0)'
 ```
 
 ### Step 6: Configure the icpos canister:
@@ -156,7 +158,7 @@ dfx deploy --network local icpos --argument '(0)'
 ic-pos uses [Courier](https://courier.com/) to send email and SMS notifications. If you want to enable notifications, you need to sign up for a Courier account and and create and API key. Then issue the following command:
 
 ```bash
-dfx canister --network local call icpos setCourierApiKey "pk_prod_..."
+dfx canister call icpos setCourierApiKey "pk_prod_..."
 ```
 
 ### Step 7: Build and run the frontend:
@@ -164,17 +166,12 @@ dfx canister --network local call icpos setCourierApiKey "pk_prod_..."
 Run npm to install dependencies and start the frontend. The frontend will be available at http://localhost:5173.
 
 ```bash
-npm install
-npm run dev
+pnpm install
+dfx deploy icpos_frontend
 ```
-
-Why don't we deploy the frontend as a local canister? Vite uses lazy loading of modules. This does not work when deploying to a local canister. When deploying to the IC mainnet, this is not an issue. Also, running using `npm run dev` allows for hot reloading of the frontend code when making changes.
-
 ### Step 8: Make a transfer!
 
 Now that everything is up and running, you can make a transfer to your newly created store.
-
-Transfers made from the owner principal will not trigger notifications in the UI since they are regarded as `mint` transactions. To test notifications, you need to make a transfer from another principal.
 
 The easiest way to do this is to create two stores using two different Internet Identity accounts, using two different web browsers. Then, transfer some tokens from one store to the other.
 
@@ -183,7 +180,7 @@ The easiest way to do this is to create two stores using two different Internet 
 Log in to the frontend using the Internet Identity. Configure the store and navigate to the `Receive` page. Click on the principal pill to copy the address to your clipboard. Then, using the `dfx` command, mint some tokens from your owner principal to the store principal.
 
 ```bash
-dfx canister --network local call icrc1_ledger icrc1_transfer '
+dfx canister call icrc1_ledger icrc1_transfer '
   (record {
     to=(record {
       owner=(principal "[STORE PRINCIPAL 1 HERE]")
@@ -205,13 +202,10 @@ If everything is working, you should see a notification in the second store.
 
 ## Possible improvements
 
-- Login state is not persisted. Reloading the app will log the user out. This should be done using `localStorage` or `sessionStorage`.
-- Show more information about transactions. A transaction detail page.
+- Show more information about transactions. 
+  - A transaction detail page.
+  - Pagination, currently only the first 5 transactions are shown.
 - Show a confirmation dialog after the user clicks on the `Send` button.
-
-## Known issues
-
--
 
 ## Contributing
