+++ motoko/ic-pos/public/cash-register.mp3
