+++ motoko/ic-pos/src/icpos/main.mo
@@ -1,20 +1,14 @@
 // Importing base modules
 import Array "mo:base/Array";
 import Blob "mo:base/Blob";
-import Char "mo:base/Char";
 import Cycles "mo:base/ExperimentalCycles";
 import Debug "mo:base/Debug";
-import HashMap "mo:base/HashMap";
-import Hash "mo:base/Hash";
 import Nat "mo:base/Nat";
-import Nat32 "mo:base/Nat32";
 import Nat64 "mo:base/Nat64";
-import Option "mo:base/Option";
 import Principal "mo:base/Principal";
 import Text "mo:base/Text";
 import Time "mo:base/Time";
 import Trie "mo:base/Trie";
-import TrieMap "mo:base/TrieMap";
 import Buffer "mo:base/Buffer";
 
 // Importing local modules
