--- motoko/ic-pos/src/icpos_frontend/canisters/ic-pos/state/merchant.state.ts
@@ -1,12 +1,0 @@
-import { Merchant } from "../../../../declarations/icpos/icpos.did";
-import { atom } from "recoil";
-
-type MerchantStateType = {
-  initialized: boolean;
-  merchant: Merchant | undefined;
-};
-
-export const MerchantState = atom<MerchantStateType>({
-  key: "MerchantState",
-  default: { initialized: false, merchant: undefined },
-});
