+++ motoko/ic-pos/src/icpos_frontend/components/HistoryButton.tsx
@@ -2,17 +2,10 @@ import { Link } from "@tanstack/react-router";
 import { Button } from "./ui/button";
 import { Inbox } from "lucide-react";
 
-type HistoryButtonProps = {
-  principal: string;
-};
-
-export default function HistoryButton({ principal }: HistoryButtonProps) {
+export default function HistoryButton() {
   return (
     <Link
       to="/history"
-      search={{
-        principal,
-      }}
       className="print:hidden"
     >
       <Button size={"lg"} className="w-56">
