+++ motoko/ic-pos/src/icpos_frontend/components/charge/key.type.ts
