+++ motoko/ic-pos/src/icpos_frontend/components/receive/TransactionOverlay.tsx
