+++ motoko/ic-pos/src/icpos_frontend/routes/login.tsx
@@ -1,7 +1,8 @@
-import { Link, createFileRoute, redirect } from '@tanstack/react-router'
+import { createFileRoute, redirect } from '@tanstack/react-router'
 import Page from '../components/Page'
 import MainSection from '../components/MainSection'
 import LoginButton from '../components/start/LoginButton'
+import ckBtc from '../assets/ckBTC.svg'
 
 export const Route = createFileRoute('/login')({
   beforeLoad: ({ context }) => {
@@ -15,21 +16,7 @@ export const Route = createFileRoute('/login')({
   component: LoginPage
 })
 
-import ckBtc from '../assets/ckBTC.svg'
-
 function LoginPage() {
-  // const { merchantState } = useIcPos();
-  //
-  // // If the merchant state is initialized it means that the backend icpos actor is available
-  // if (merchantState.initialized) {
-  //   // If the merchant is initialized, navigate to the merchant page
-  //   if (merchantState.merchant) {
-  //     return <Navigate to="/merchant" />;
-  //   }
-  //   // Otherwise, navigate to the config page to create a merchant
-  //   return <Navigate to="/initial-config" />;
-  // }
-
   return (
     <Page>
       <MainSection>
@@ -40,11 +27,7 @@ function LoginPage() {
           <div className="text-4xl font-bold">IC-POS </div>
           <div className="text-center">
             Setup a simple store front to accept ckBTC payments on the Internet
-            Computer. Sign in with Internet Identity to get started or{' '}
-            <Link to="/receive-select-principal">
-              monitor any address without signing in
-            </Link>
-            .
+            Computer. Sign in with Internet Identity to get started.
           </div>
           <div className="grow" />
           <LoginButton />
